Provide "stock" and "management_unit_name" options for a run reconstruction dataframe

Once a run reconstruction dataframe is created, it's nice to see the options for stock and management_unit_name identifiers for making plots. This function acts as a wrapper for unique(data$stock) if no stock name is provided; if stock name is provided, instead is a wrapper for filtering to that stock and then unique(data$management_unit_name. When presenting management unit names, removes astrixes; this is consistent with plot_rr_mu() behavior.

Usage

list_rr_options(data, stock = NULL)

Arguments

data

Dataframe generated by read_rr_trs()

stock

Optional, stock

Value

character vector