Plot timeseries of values for a given management unit.

Plot timeseries of values for a given management unit.

Usage

plot_rr_mu(
  data,
  stock.use,
  mu.use,
  normalize_fisheries = TRUE,
  base_font_size = 13
)

Arguments

data

Dataframe generated from read_rr_trs()

stock.use

Character string of stock name (which identify blocks within the run reconstruction excel file) to plot. Use list_rr_options() to see options.

mu.use

Character string of "management_unit_name" (which identify rows within each block of the run reconstruction excel file) to plot. Use list_rr_options() to see options. Matches versions of management unit name with and without astrices, as those are only used to add notes and should not be treated as new versions of a management unit.

normalize_fisheries

Logical: should the calues of fisheries by turned to % of this stock's catch per fishery in each year (TRUE) or presented as raw values (FALSE). Defaults to TRUE.

base_font_size

Integer passed to theme_bw() to set baseline font size in figures. Defaults to 13.

Value

Figure object made of ggplots combined using {patchwork}.